Move KuduScanEnumerator to new file
